+++ src/passes/DeadArgumentElimination.cpp
@@ -210,10 +210,8 @@ struct DAEScanner
     std::vector<Item> work;
     work.emplace_back(entry, initial);
     while (!work.empty()) {
-      auto item = std::move(work.back());
+      auto [block, indexes] = std::move(work.back());
       work.pop_back();
-      auto* block = item.first;
-      auto& indexes = item.second;
       // Ignore things we've already seen, or we've already seen to be used.
       auto& seenIndexes = seenBlockIndexes[block];
       indexes.filter([&](const Index i) {
@@ -295,19 +293,16 @@ struct DAE : public Pass {
     // Combine all the info.
     std::unordered_map<Name, std::vector<Call*>> allCalls;
     std::unordered_set<Name> tailCallees;
-    for (auto& pair : infoMap) {
-      auto& info = pair.second;
-      for (auto& pair : info.calls) {
-        auto name = pair.first;
-        auto& calls = pair.second;
+    for (auto& [_, info] : infoMap) {
+      for (auto& [name, calls] : info.calls) {
         auto& allCallsToName = allCalls[name];
         allCallsToName.insert(allCallsToName.end(), calls.begin(), calls.end());
       }
       for (auto& callee : info.tailCallees) {
         tailCallees.insert(callee);
       }
-      for (auto& pair : info.droppedCalls) {
-        allDroppedCalls[pair.first] = pair.second;
+      for (auto& [name, calls] : info.droppedCalls) {
+        allDroppedCalls[name] = calls;
       }
     }
     // If we refine return types then we will need to do more type updating
@@ -315,13 +310,11 @@ struct DAE : public Pass {
     bool refinedReturnTypes = false;
     // We now have a mapping of all call sites for each function, and can look
     // for optimization opportunities.
-    for (auto& pair : allCalls) {
-      auto name = pair.first;
+    for (auto& [name, calls] : allCalls) {
       // We can only optimize if we see all the calls and can modify them.
       if (infoMap[name].hasUnseenCalls) {
         continue;
       }
-      auto& calls = pair.second;
       auto* func = module->getFunction(name);
       auto numParams = func->getNumParams();
       // Refine argument types before doing anything else. This does not
@@ -376,9 +369,7 @@ struct DAE : public Pass {
     // Track which functions we changed, and optimize them later if necessary.
     std::unordered_set<Function*> changed;
     // We now know which parameters are unused, and can potentially remove them.
-    for (auto& pair : allCalls) {
-      auto name = pair.first;
-      auto& calls = pair.second;
+    for (auto& [name, calls] : allCalls) {
       if (infoMap[name].hasUnseenCalls) {
         continue;
       }
@@ -553,6 +544,7 @@ private:
     auto numParams = func->getNumParams();
     std::vector<Type> newParamTypes;
     newParamTypes.reserve(numParams);
+    std::vector<LUBFinder> lubs(numParams);
     for (Index i = 0; i < numParams; i++) {
       auto originalType = func->getLocalType(i);
       // If the parameter type is not a reference, there is nothing to refine.
@@ -565,10 +557,11 @@ private:
         newParamTypes.push_back(originalType);
         continue;
       }
-      LUBFinder lub;
+      auto& lub = lubs[i];
       for (auto* call : calls) {
         auto* operand = call->operands[i];
-        if (lub.note(operand) == originalType) {
+        lub.noteUpdatableExpression(operand);
+        if (lub.getBestPossible() == originalType) {
           // We failed to refine this parameter to anything more specific.
           break;
         }
@@ -578,7 +571,7 @@ private:
       if (!lub.noted()) {
         return;
       }
-      newParamTypes.push_back(lub.get());
+      newParamTypes.push_back(lub.getBestPossible());
     }
 
     // Check if we are able to optimize here before we do the work to scan the
@@ -590,6 +583,11 @@ private:
 
     // We can do this!
     TypeUpdating::updateParamTypes(func, newParamTypes, *module);
+
+    // Update anything the lubs need to update.
+    for (auto& lub : lubs) {
+      lub.updateNulls();
+    }
 
     // Also update the function's type.
     func->setParams(newParams);
@@ -633,21 +631,28 @@ private:
     ReFinalize().walkFunctionInModule(func, module);
 
     LUBFinder lub;
-    if (lub.note(func->body) == originalType) {
+    lub.noteUpdatableExpression(func->body);
+    if (lub.getBestPossible() == originalType) {
       return false;
     }
 
-    // Scan the body and look at the returns.
-    auto processReturnType = [&](Type type) {
-      // Return whether we still look ok to do the optimization. If this is
-      // false then we can stop here.
-      return lub.note(type) != originalType;
-    };
+    // Scan the body and look at the returns. First, return expressions.
     for (auto* ret : FindAll<Return>(func->body).list) {
-      if (!processReturnType(ret->value->type)) {
+      lub.noteUpdatableExpression(ret->value);
+      if (lub.getBestPossible() == originalType) {
         return false;
       }
     }
+
+    // Process return_calls and call_refs. Unlike return expressions which we
+    // just handled, these only get a type to update, not a value.
+    auto processReturnType = [&](Type type) {
+      // Return whether we still look ok to do the optimization. If this is
+      // false then we can stop here.
+      lub.note(type);
+      return lub.getBestPossible() != originalType;
+    };
+
     for (auto* call : FindAll<Call>(func->body).list) {
       if (call->isReturn &&
           !processReturnType(module->getFunction(call->target)->getResults())) {
@@ -655,7 +660,8 @@ private:
       }
     }
     for (auto* call : FindAll<CallIndirect>(func->body).list) {
-      if (call->isReturn && !processReturnType(call->sig.results)) {
+      if (call->isReturn &&
+          !processReturnType(call->heapType.getSignature().results)) {
         return false;
       }
     }
@@ -671,7 +677,6 @@ private:
         }
       }
     }
-    assert(lub.get() != originalType);
 
     // If the refined type is unreachable then nothing actually returns from
     // this function.
@@ -680,13 +685,19 @@ private:
       return false;
     }
 
+    auto newType = lub.getBestPossible();
+    if (newType == originalType) {
+      return false;
+    }
+
     // Success. Update the type, and the calls.
-    func->setResults(lub.get());
+    func->setResults(newType);
     for (auto* call : calls) {
       if (call->type != Type::unreachable) {
-        call->type = lub.get();
+        call->type = newType;
       }
     }
+    lub.updateNulls();
     return true;
   }
 };