+++ src/passes/OptimizeInstructions.cpp
@@ -1129,7 +1129,7 @@ struct OptimizeInstructions
         // Otherwise, if this is not a tee, then no value falls through. The
         // ref.as_non_null acts as a null check here, basically. If we are
         // ignoring such traps, we can remove it.
-        auto passOptions = getPassOptions();
+        auto& passOptions = getPassOptions();
         if (passOptions.ignoreImplicitTraps || passOptions.trapsNeverHappen) {
           curr->value = as->value;
         }
@@ -1253,7 +1253,7 @@ struct OptimizeInstructions
                        .makeCallIndirect(get->table,
                                          get->index,
                                          curr->operands,
-                                         get->type.getHeapType().getSignature(),
+                                         get->type.getHeapType(),
                                          curr->isReturn));
       return;
     }
@@ -1540,7 +1540,7 @@ struct OptimizeInstructions
     }
 
     Builder builder(*getModule());
-    auto passOptions = getPassOptions();
+    auto& passOptions = getPassOptions();
 
     auto fallthrough =
       Properties::getFallthrough(curr->ref, getPassOptions(), *getModule());
@@ -1869,7 +1869,7 @@ private:
     // assume things like local.get's of the same index being identical. (It is
     // also ok to have side effects here, if we can remove them, as we are also
     // checking if we can remove the two inputs anyhow.)
-    auto passOptions = getPassOptions();
+    auto& passOptions = getPassOptions();
     if (EffectAnalyzer(passOptions, *getModule(), left)
           .hasUnremovableSideEffects() ||
         EffectAnalyzer(passOptions, *getModule(), right)
@@ -2558,21 +2558,59 @@ private:
       }
     }
     {
-      // (i32(x) != 0) | (i32(y) != 0)   ==>   i32(x | y) != 0
-      // (i64(x) != 0) | (i64(y) != 0)   ==>   i64(x | y) != 0
+      // Binary operations that preserve a bitwise OR can be
+      // reordered. If F(x) = binary(x, c), and F(x) preserves OR,
+      // that is,
+      //
+      //   F(x) | F(y) == F(x | y)
+      //
+      // Then also
+      //
+      //   binary(x, c) | binary(y, c)  =>  binary(x | y, c)
+      Binary *bx, *by;
       Expression *x, *y;
+      Const *cx, *cy;
       if (matches(curr,
                   binary(OrInt32,
-                         binary(Ne, any(&x), ival(0)),
-                         binary(Ne, any(&y), ival(0)))) &&
-          x->type == y->type) {
-        auto* inner = curr->left->cast<Binary>();
-        inner->left = Builder(*getModule())
-                        .makeBinary(Abstract::getBinary(x->type, Or), x, y);
-        return inner;
+                         binary(&bx, any(&x), ival(&cx)),
+                         binary(&by, any(&y), ival(&cy)))) &&
+          bx->op == by->op && x->type == y->type && cx->value == cy->value &&
+          preserveOr(bx)) {
+        bx->left = Builder(*getModule())
+                     .makeBinary(Abstract::getBinary(x->type, Or), x, y);
+        return bx;
       }
     }
     return nullptr;
+  }
+
+  // Check whether an operation preserves the Or operation through it, that is,
+  //
+  //   F(x | y) = F(x) | F(y)
+  //
+  // Mathematically that means F is homomorphic with respect to the | operation.
+  //
+  // F(x) is seen as taking a single parameter of its first child. That is, the
+  // first child is |x|, and the rest is constant. For example, if we are given
+  // a binary with operation != and the right child is a constant 0, then
+  // F(x) = (x != 0).
+  bool preserveOr(Binary* curr) {
+    using namespace Abstract;
+    using namespace Match;
+
+    // (x != 0) | (y != 0)    ==>    (x | y) != 0
+    // This effectively checks if any bits are set in x or y.
+    if (matches(curr, binary(Ne, any(), ival(0)))) {
+      return true;
+    }
+
+    // (x < 0) | (y < 0)    ==>    (x | y) < 0
+    // This effectively checks if x or y have the sign bit set.
+    if (matches(curr, binary(LtS, any(), ival(0)))) {
+      return true;
+    }
+
+    return false;
   }
 
   // fold constant factors into the offset
@@ -3307,7 +3345,7 @@ private:
   }
 
   Expression* optimizeMemoryCopy(MemoryCopy* memCopy) {
-    PassOptions options = getPassOptions();
+    auto& options = getPassOptions();
 
     if (options.ignoreImplicitTraps || options.trapsNeverHappen) {
       if (ExpressionAnalyzer::equal(memCopy->dest, memCopy->source)) {
@@ -3386,7 +3424,7 @@ private:
       return nullptr;
     }
 
-    PassOptions options = getPassOptions();
+    auto& options = getPassOptions();
     Builder builder(*getModule());
 
     auto* csize = memFill->size->cast<Const>();